Add dropout_to_identity transform and related flag in quantize_pact
This is useful when deploying networks including Dropout layers,
which are not useful in inference and cause issues in the DeployGraph.
Note that the Dropout removal has to be done in FullPrecision stage,
i.e., before creating the deploy graph: this commits adds a flag
to quantize_pact in order to do this (should probably not be active
when doing retraining).