Prune useless BN parameters after Linear weight hardening
Batch-Normalization parameter quantization is easier when the
parameters do not have outliers (as quantization is performed
per tensor). These outliers are common when the previous Linear
layer's weights are small along the same channel - a condition
that often results in them being always 0 after hardening.
To ease BN quantization, we introduce a transformation that
prunes BN parameters when they follow a channel with 0'ed weights
(i.e., it is effectively unused). This is active by default when
switching to QD stage.