Update pact.py
Corrected the forward funztion  of class PACT_IntegerAdd(torch.nn.Module) to handle  the following case:
if the PACT_IntegerAdd doesnt have the "eps_out" filed, it checks if "eps_in_list" attribute is available. In this case it calculates the "eps_out". 
This is added to handle the specific pulp dronet case, that doesnt have any relu after the last add module, resulting in not calculating the eps_out value for this layer